requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
qrcode[pil]==8.2
Pillow>=9.1.0
Flask==2.3.2
Flask-Session==0.5.0
user-agents==2.2.0
requests==2.31.0
beautifulsoup4==4.12.0
python-whois==0.8.0
exifread==3.0.0
python-nmap==0.7.1
